+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning].
 
 ### Fixed
 
+- codegen: Fixed invalid `models.py` template.
+- context: Do not wrap known exceptions with `CallbackError`
+- hasura: Fixed processing M2M relations.
+- sentry: Fixed "invalid value `environment`" error.
 - sentry: Ignore events from project callbacks when `crash_reporting` is enabled.
 
 ## [6.0.0] - 2022-08-08
@@ -699,7 +703,7 @@ This release contains no changes except for the version number.
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
-[unreleased]: https://github.com/dipdup-net/dipdup-py/compare/6.0.0...HEAD
+[Unreleased]: https://github.com/dipdup-net/dipdup-py/compare/6.0.0...HEAD
 [6.0.0]: https://github.com/dipdup-net/dipdup-py/compare/6.0.0rc2...6.0.0
 [6.0.0rc2]: https://github.com/dipdup-net/dipdup-py/compare/6.0.0-rc1...6.0.0rc2
 [6.0.0-rc1]: https://github.com/dipdup-net/dipdup-py/compare/5.2.5...6.0.0-rc1